+++ dashboard/src/utils/testUtils.tsx
@@ -31,6 +31,19 @@ const queryClient = new QueryClient({
   },
 });
 
+process.env = {
+  NODE_ENV: 'development',
+  NEXT_PUBLIC_NHOST_PLATFORM: 'false',
+  NEXT_PUBLIC_ENV: 'dev',
+  NEXT_PUBLIC_NHOST_AUTH_URL: 'https://localdev.nhost.run/v1/auth',
+  NEXT_PUBLIC_NHOST_FUNCTIONS_URL: 'https://localdev.nhost.run/v1/functions',
+  NEXT_PUBLIC_NHOST_GRAPHQL_URL: 'https://localdev.nhost.run/v1/graphql',
+  NEXT_PUBLIC_NHOST_STORAGE_URL: 'https://localdev.nhost.run/v1/storage',
+  NEXT_PUBLIC_NHOST_HASURA_CONSOLE_URL: 'http://localhost:9695',
+  NEXT_PUBLIC_NHOST_HASURA_MIGRATIONS_API_URL: 'http://localhost:9693',
+  NEXT_PUBLIC_NHOST_HASURA_API_URL: 'http://localhost:8080',
+};
+
 export const mockRouter: NextRouter = {
   basePath: '',
   pathname: '/',