+++ integrations/react-urql/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @nhost/react-urql
 
+## 2.0.9
+
+### Patch Changes
+
+- Updated dependencies [f375eacc]
+  - @nhost/react@2.0.9
+
+## 2.0.8
+
+### Patch Changes
+
+- @nhost/react@2.0.8
+
+## 2.0.7
+
+### Patch Changes
+
+- 850a049c: chore(deps): update docker/build-push-action action to v4
+- Updated dependencies [850a049c]
+  - @nhost/react@2.0.7
+
 ## 2.0.6
 
 ### Patch Changes