+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @nhost/vue
 
+## 1.13.15
+
+### Patch Changes
+
+- Updated dependencies [088584e7]
+  - @nhost/nhost-js@2.0.9
+
+## 1.13.14
+
+### Patch Changes
+
+- Updated dependencies [1d155559]
+  - @nhost/nhost-js@2.0.8
+
+## 1.13.13
+
+### Patch Changes
+
+- 7eac17a1: chore(deps): bump `@xstate/inspect` to 0.8.0
+- 850a049c: chore(deps): update docker/build-push-action action to v4
+- Updated dependencies [850a049c]
+  - @nhost/nhost-js@2.0.7
+
 ## 1.13.12
 
 ### Patch Changes